+++ test/qbits/alia/test/core.clj
@@ -8,6 +8,7 @@
    [qbits.alia.codec :refer :all]
    [qbits.alia.codec.joda-time :refer :all]
    [qbits.alia.codec.eaio-uuid :refer :all]
+   [qbits.alia.codec.nippy :refer :all]
    [qbits.tardis :refer :all]
    [qbits.hayt :refer :all]
    [clojure.core.async :as async]